MA-10625 fix alarm notification issue
It have below error when alarm time trigged.
There is no UI or notification when the alarm is trigged.
Error log:
 E NotificationService: No Channel found for pkg=com.android.deskclock, channelId=null,
 id=3, tag=null, opPkg=com.android.deskclock, callingUid=10043, userId=0,
 incomingUserId=0, notificationUid=10043, notification=Notification

When the target is Android O, system must implement one or more notification channels to
display notifications to users.
This patch create alarm default channel when ACTION_BOOT_COMPLETED in AlarmInitReceiver.
It need add setChannelId when create the notification in the destclock apk.

Change-Id: Ifc07b503034b232530f7f39ef9f5d17c3dd63342
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/3408
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>